prefactoring: introduce core.NameValueResult
